+++ src/components/date_picker/super_date_picker/quick_select_popover/quick_select.js
@@ -1,0 +1,194 @@
+
+import PropTypes from 'prop-types';
+import React, { Component, Fragment } from 'react';
+import moment from 'moment';
+import dateMath from '@elastic/datemath';
+
+import { EuiButton, EuiButtonIcon } from '../../../button';
+import { EuiFlexGroup, EuiFlexItem } from '../../../flex';
+import { EuiTitle } from '../../../title';
+import { EuiSpacer } from '../../../spacer';
+import { EuiFormRow, EuiSelect, EuiFieldNumber } from '../../../form';
+import { EuiToolTip } from '../../../tool_tip';
+import { EuiHorizontalRule } from '../../../horizontal_rule';
+
+import { timeUnits } from '../time_units';
+
+const LAST = 'last';
+const NEXT = 'next';
+
+const timeTenseOptions = [
+  { value: LAST, text: 'Last' },
+  { value: NEXT, text: 'Next' },
+];
+const timeUnitsOptions = Object.keys(timeUnits).map(key => {
+  return { value: key, text: `${timeUnits[key]}s` };
+});
+
+export class EuiQuickSelect extends Component {
+  constructor(props) {
+    super(props);
+
+    const { timeTense, timeValue, timeUnits } = this.props.prevQuickSelect;
+    this.state = {
+      timeTense: timeTense ? timeTense : LAST,
+      timeValue: timeValue ? timeValue : 15,
+      timeUnits: timeUnits ? timeUnits : 'm',
+    };
+  }
+
+  onTimeTenseChange = (evt) => {
+    this.setState({
+      timeTense: evt.target.value,
+    });
+  }
+
+  onTimeValueChange = (evt) => {
+    const sanitizedValue = parseInt(evt.target.value, 10);
+    this.setState({
+      timeValue: isNaN(sanitizedValue) ? '' : sanitizedValue,
+    });
+  }
+
+  onTimeUnitsChange = (evt) => {
+    this.setState({
+      timeUnits: evt.target.value,
+    });
+  }
+
+  applyQuickSelect = () => {
+    const {
+      timeTense,
+      timeValue,
+      timeUnits,
+    } = this.state;
+
+    if (timeTense === NEXT) {
+      this.props.applyTime({
+        start: 'now',
+        end: `now+${timeValue}${timeUnits}`,
+        quickSelect: { ...this.state },
+      });
+      return;
+    }
+
+    this.props.applyTime({
+      start: `now-${timeValue}${timeUnits}`,
+      end: 'now',
+      quickSelect: { ...this.state },
+    });
+  }
+
+  getBounds = () => {
+    return {
+      min: dateMath.parse(this.props.start),
+      max: dateMath.parse(this.props.end, { roundUp: true }),
+    };
+  }
+
+  stepForward = () => {
+    const { min, max } = this.getBounds();
+    const diff = max.diff(min);
+    this.props.applyTime({
+      start: moment(max).add(1, 'ms').toISOString(),
+      end: moment(max).add(diff + 1, 'ms').toISOString(),
+      keepPopoverOpen: true,
+    });
+  }
+
+  stepBackward = () => {
+    const { min, max } = this.getBounds();
+    const diff = max.diff(min);
+    this.props.applyTime({
+      start: moment(min).subtract(diff + 1, 'ms').toISOString(),
+      end: moment(min).subtract(1, 'ms').toISOString(),
+      keepPopoverOpen: true,
+    });
+  }
+
+  render() {
+    return (
+      <Fragment>
+        <EuiFlexGroup responsive={false} alignItems="center" gutterSize="s">
+          <EuiFlexItem>
+            <EuiTitle size="xxxs"><span>Quick select</span></EuiTitle>
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiToolTip content="Previous time window">
+              <EuiButtonIcon
+                aria-label="Previous time window"
+                iconType="arrowLeft"
+                onClick={this.stepBackward}
+              />
+            </EuiToolTip>
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiToolTip content="Next time window">
+              <EuiButtonIcon
+                aria-label="Next time window"
+                iconType="arrowRight"
+                onClick={this.stepForward}
+              />
+            </EuiToolTip>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+        <EuiSpacer size="s" />
+        <EuiFlexGroup gutterSize="s" responsive={false}>
+          <EuiFlexItem>
+            <EuiFormRow compressed>
+              <EuiSelect
+                aria-label="Quick time tense"
+                value={this.state.timeTense}
+                options={timeTenseOptions}
+                onChange={this.onTimeTenseChange}
+              />
+            </EuiFormRow>
+          </EuiFlexItem>
+          <EuiFlexItem>
+            <EuiFormRow compressed>
+              <EuiFieldNumber
+                aria-label="Quick time value"
+                value={this.state.timeValue}
+                onChange={this.onTimeValueChange}
+              />
+            </EuiFormRow>
+          </EuiFlexItem>
+          <EuiFlexItem>
+            <EuiFormRow compressed>
+              <EuiSelect
+                aria-label="Quick time units"
+                value={this.state.timeUnits}
+                options={timeUnitsOptions}
+                onChange={this.onTimeUnitsChange}
+              />
+            </EuiFormRow>
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiFormRow>
+              <EuiButton
+                size="s"
+                onClick={this.applyQuickSelect}
+                style={{ minWidth: 0 }}
+                disabled={this.state.timeValue === '' || this.state.timeValue <= 0}
+              >
+                Apply
+              </EuiButton>
+            </EuiFormRow>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+        <EuiHorizontalRule margin="s" />
+      </Fragment>
+    );
+  }
+}
+
+EuiQuickSelect.propTypes = {
+  applyTime: PropTypes.func.isRequired,
+  start: PropTypes.string.isRequired,
+  end: PropTypes.string.isRequired,
+  prevQuickSelect: PropTypes.object,
+};
+
+EuiQuickSelect.defaultProps = {
+  prevQuickSelect: {},
+};